fix(showcase,core,adapters): unlayered utilities + Sonar back to zero
The demo's Filters/Columns buttons rendered with no padding or background.
Root cause (same disease as the earlier border kill): the kit demos inject
unlayered global resets — Chakra's CSSReset sets `button { padding: 0;
background: transparent }` — and unlayered CSS beats every @layer-ed rule,
so the demo's px-*/bg-* utilities silently lost. Categorical fix: Tailwind
utilities now import UNLAYERED, so a class (0,1,0) outranks any kit's
element reset; the old per-property border patches are gone. Verified by
computed styles in light + dark: padding 0 12px, 1px border, 6px radius,
correct backgrounds — pills and popovers intact.

Sonar findings from the declarative branch, all fixed (scorecard back to
100.0 coverage / 0 bugs / 0 vulnerabilities / 0 smells / 0 hotspots /
0.0% duplication, gate OK):
- the column-menu labels/props contract was declared verbatim in five
  adapters — hoisted into core (ColumnMenuLabels, ColumnMenuChromeProps)
  and adopted everywhere
- the unstyled rows-per-page select was duplicated between toolbar and
  footer — extracted as a shared RowsPerPageSelect
- the refresh indicator is a native <progress> (implicit progressbar role)
- a double import in columnMenuModel merged
