+++ packages/core/src/columns/columnMenuModel.ts
@@ -1,5 +1,5 @@
 import type { ColumnDef } from "../types";
-import type { UseColumnLayoutResult } from "./useColumnLayout";
+import type { PinSide, UseColumnLayoutResult } from "./useColumnLayout";
 import { applyColumnOrder } from "./useColumnLayout";
 
 /** Readable label for a column in the menu (header string → mobileLabel → key). */
@@ -7,8 +7,6 @@ export function columnMenuLabel<TRow>(column: ColumnDef<TRow>): string {
   if (typeof column.header === "string") return column.header;
   return column.mobileLabel ?? column.key;
 }
-
-import type { PinSide } from "./useColumnLayout";
 
 /** Edge a column is pinned to, or `undefined` when unpinned. */
 export type PinnedSide = PinSide | undefined;
@@ -65,4 +63,31 @@ export function columnMenuRows<TRow>(
       index,
     })
   );
+}
+
+/**
+ * Labels every adapter's column menu needs (pre-translated by the caller).
+ * Hoisted here so the five adapters share one contract instead of
+ * re-declaring it.
+ */
+export interface ColumnMenuLabels {
+  columns: string;
+  pinLeft: string;
+  pinRight: string;
+  unpin: string;
+  moveLeft: string;
+  moveRight: string;
+  resetColumns: string;
+  showColumn: string;
+  hideColumn: string;
+}
+
+/** The shared prop surface of every adapter's `<ColumnMenu>`. */
+export interface ColumnMenuChromeProps<TRow> {
+  /** All declared columns (pre layout filtering). */
+  allColumns: ColumnDef<TRow>[];
+  /** The user column-layout state + mutators. */
+  layout: UseColumnLayoutResult<TRow>;
+  /** Resolved labels. */
+  labels: ColumnMenuLabels;
 }
